backend/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pydantic
pydantic-settings

sqlalchemy
passlib[bcrypt]
python-jose

pymupdf
numpy
faiss-cpu
sentence-transformers
rank-bm25
httpx
openai
